Refactor(test): move testAcceptsTimeParameter to ParamsTests
See https://github.com/GIScience/openrouteservice/pull/1376
